requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-pinecone
faiss-cpu
pypdf
python-dotenv
numpy
pinecone-client
